Format account and child names consistently in employee desktop and mobile
