+++ frontend/src/lib-common/messaging.ts
@@ -4,10 +4,14 @@
 
 import {
   DraftContent,
+  MessageAccount,
+  MessageChild,
   SentMessage
 } from 'lib-common/generated/api-types/messaging'
 import HelsinkiDateTime from 'lib-common/helsinki-date-time'
 import { JsonOf } from 'lib-common/json'
+
+import { formatFirstName } from './names'
 
 export const deserializeDraftContent = ({
   created,
@@ -24,3 +28,23 @@ export const deserializeSentMessage = ({
   ...rest,
   sentAt: HelsinkiDateTime.parseIso(sentAt)
 })
+
+export function formatAccountNames(
+  sender: MessageAccount,
+  recipients: MessageAccount[],
+  children: MessageChild[]
+): { senderName: string; recipientNames: string[] } {
+  const childNames =
+    children.length > 0
+      ? children.map((child) => formatFirstName(child)).join(', ')
+      : undefined
+  const childSuffix = childNames ? ` (${childNames})` : ''
+
+  const senderName =
+    sender.name + (sender.type === 'CITIZEN' ? childSuffix : '')
+  const recipientNames = recipients.map(
+    (r) => r.name + (r.type === 'CITIZEN' ? childSuffix : '')
+  )
+
+  return { senderName, recipientNames }
+}
